[OV] Align FQ params data type with target node  precision (#1585)
### Changes

For models coming from a framework in FP16 precision, FakeQuantize
parameters should be in FP16.
For models coming from a framework in FP32 precision and converted to
FP16 via MO, FakeQuantize parameters should be in FP32.
FQ parameters precision aligned with model input precision.
Input FP32

![image](https://user-images.githubusercontent.com/22346860/221900280-a4e0746c-403d-4e36-8e1c-d3103f91ba58.png)
Input FP16

![image](https://user-images.githubusercontent.com/22346860/221900547-88280cfc-6afa-4625-b7ef-e76631d35ee5.png)


### Reason for changes

* To align NNCF PTQ with POT

### Related tickets

* 103926, 101738

### Tests

<!--- How was the correctness of changes tested and whether new tests
were added -->

+++ nncf/experimental/openvino_native/quantization/quantizer_parameters.py
@@ -204,8 +204,8 @@ def calculate_quantizer_parameters(statistics: MinMaxTensorStatistic,
     :param quantizer_group: Group of the quantizer.
     :return: Parameters of the FakeQuantize layer.
     """
-    min_values = np.array(statistics.min_values)
-    max_values = np.array(statistics.max_values)
+    min_values = np.array(statistics.min_values).astype(np.float32)
+    max_values = np.array(statistics.max_values).astype(np.float32)
 
     if quantizer_config.mode == QuantizationMode.SYMMETRIC:
         narrow_range = quant_group == QuantizerGroup.WEIGHTS
